make r int

+++ Lattices/Lattices/Lattice.cpp
@@ -192,7 +192,7 @@ void Lattice::LLLalgorithm()
 		//step 3
 		if (abs(mu[k][k-1]) > 0.5)
 		{
-			double r = mu[k][k-1] > 0 ? int(0.5 + mu[k][k-1]) : -int(0.5-mu[k][k-1]);
+			int r = mu[k][k-1] > 0 ? int(0.5 + mu[k][k-1]) : -int(0.5-mu[k][k-1]);
 			for (int i = 0; i < _size; i++)
 				_basis[k][i] -= r*_basis[k-1][i];
 			for (int j = 0; j < k-2; j++)
@@ -237,7 +237,7 @@ void Lattice::LLLalgorithm()
 			{
 				if (abs(mu[k][l]) > 0.5)
 				{
-					double r = mu[k][l] > 0 ? int(0.5 + mu[k][l]) : -int(0.5-mu[k][l]);
+					int r = mu[k][l] > 0 ? int(0.5 + mu[k][l]) : -int(0.5-mu[k][l]);
 					for (int i = 0; i < _size; i++)
 						_basis[k][i] -= r*_basis[l][i];
 					for (int j = 0; j < l-1; j++)
